# feat(Query): Add random keyword in DQL