Fixes: https://discuss.dgraph.io/t/pick-random-node/10815

## Changes

 - random nodes can be selected from the graph using `random: <number of nodes>`
 - random selection by replacement is used to select nodes

## Issues/future work

 - Since selection is by replacement, you can get duplicates
 - No error handling yet in the `applyRandom` function
 - Making of new slice in lines 2430-2432 looks a little inefficient, there might a better way to do this.

This is my first PR here (and I'm new to Go), there will be a lot more issues I'm not aware of.

## Example

I have only tested this on a basic example from Ratel. This will need a lot more testing.

<details>
  <summary>Mutation</summary>

  ```graphql
  {
    set {
     _:luke <name> "Luke Skywalker" .
     _:luke <dgraph.type> "Person" .
     _:leia <name> "Princess Leia" .
     _:leia <dgraph.type> "Person" .
     _:han <name> "Han Solo" .
     _:han <dgraph.type> "Person" .
     _:lucas <name> "George Lucas" .
     _:lucas <dgraph.type> "Person" .
     _:irvin <name> "Irvin Kernshner" .
     _:irvin <dgraph.type> "Person" .
     _:richard <name> "Richard Marquand" .
     _:richard <dgraph.type> "Person" .
  
     _:sw1 <name> "Star Wars: Episode IV - A New Hope" .
     _:sw1 <release_date> "1977-05-25" .
     _:sw1 <revenue> "775000000" .
     _:sw1 <running_time> "121" .
     _:sw1 <starring> _:luke .
     _:sw1 <starring> _:leia .
     _:sw1 <starring> _:han .
     _:sw1 <director> _:lucas .
     _:sw1 <dgraph.type> "Film" .
  
     _:sw2 <name> "Star Wars: Episode V - The Empire Strikes Back" .
     _:sw2 <release_date> "1980-05-21" .
     _:sw2 <revenue> "534000000" .
     _:sw2 <running_time> "124" .
     _:sw2 <starring> _:luke .
     _:sw2 <starring> _:leia .
     _:sw2 <starring> _:han .
     _:sw2 <director> _:irvin .
     _:sw2 <dgraph.type> "Film" .
  
     _:sw3 <name> "Star Wars: Episode VI - Return of the Jedi" .
     _:sw3 <release_date> "1983-05-25" .
     _:sw3 <revenue> "572000000" .
     _:sw3 <running_time> "131" .
     _:sw3 <starring> _:luke .
     _:sw3 <starring> _:leia .
     _:sw3 <starring> _:han .
     _:sw3 <director> _:richard .
     _:sw3 <dgraph.type> "Film" .
  
     _:st1 <name> "Star Trek: The Motion Picture" .
     _:st1 <release_date> "1979-12-07" .
     _:st1 <revenue> "139000000" .
     _:st1 <running_time> "132" .
     _:st1 <dgraph.type> "Film" .
    }
}
  ```
</details>

**Query**

```graphql
{
  q(func: type(Person), random: 5) {
    name
    uid
  }
}
```

**Result**

```json
  "data": {
    "q": [
      {
        "name": "Luke Skywalker",
        "uid": "0x7538"
      },
      {
        "name": "Luke Skywalker",
        "uid": "0x7538"
      },
      {
        "name": "George Lucas",
        "uid": "0x7539"
      },
      {
        "name": "Princess Leia",
        "uid": "0x753c"
      },
      {
        "name": "Richard Marquand",
        "uid": "0x7535"
      }
    ]
  }
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/dgraph/7693)
<!-- Reviewable:end -->