@Samyak2 2021-04-22
Yes, I'm still working on it. Sorry, I couldn't do it in the last week because of the end sem workload from college 😄.
I will implement your suggestions by tonight (IST).