@Samyak2 2021-04-22
@ahsanbarkati thank you for the explanation, it makes a lot more sense now.

I have implemented your suggestions in the latest commit:
 - `k` random uids are now selected from each uid list instead of the whole matrix
 - for in-place selection, I have first shuffled the list (in-place) using [`rand.Shuffle`](https://golang.org/pkg/math/rand/#Shuffle) and then selected the first `k` elements. Would it be better to just select a random window here instead of the shuffling the list? Are uids guaranteed to be ordered randomly in the list?

I will look at adding tests for these now.