@Samyak2 2021-04-25
I had a few issues while running the tests:

 - The contributing guidelines instruct to [run `test.sh`](https://github.com/dgraph-io/dgraph/blob/master/CONTRIBUTING.md#testing) whereas the script is apparently deprecated and `t/t.go` is to be used instead. Should this be updated in the guidelines?
 - I get an error while running the tests with `./t` or `./t -s -p "query"`

<details>
  <summary>Output of `./t`</summary>
  
  ```log
testutil: "" localhost:9080 localhost:5080
Proc ID is 992
Installing dgraph ...
make[1]: Entering directory '/home/samyak/PESU/Sem6/HP/Project/dgraph/dgraph'
Commit SHA256: a44abf8587f4c26714952c45a05aedff1373ab13
Old SHA256: 561df6fcce9beff32df455feed232cf4b9d1a726fed7a4ca93abf44613ab7260
New SHA256: 561df6fcce9beff32df455feed232cf4b9d1a726fed7a4ca93abf44613ab7260
make[1]: Leaving directory '/home/samyak/PESU/Sem6/HP/Project/dgraph/dgraph'
compiling plugin: src="./custom_plugins/anagram/main.go" so="./custom_plugins/0.so"
compiling plugin: src="./custom_plugins/cidr/main.go" so="./custom_plugins/1.so"
compiling plugin: src="./custom_plugins/factor/main.go" so="./custom_plugins/2.so"
compiling plugin: src="./custom_plugins/rune/main.go" so="./custom_plugins/3.so"
plugin build completed. Files are: /home/samyak/PESU/Sem6/HP/Project/dgraph/t/custom_plugins/0.so,/home/samyak/PESU/Sem6/HP/Project/dgraph/t/custom_plugins/1.so,/home/samyak/PESU/Sem6/HP/Project/dgraph/t/custom_plugins/2.so,/home/samyak/PESU/Sem6/HP/Project/dgraph/t/custom_plugins/3.so
Skipping package github.com/dgraph-io/dgraph/systest/1million as its not valid for the selected suite unit 
Skipping package github.com/dgraph-io/dgraph/systest/21million/bulk as its not valid for the selected suite unit 
Skipping package github.com/dgraph-io/dgraph/systest/21million/live as its not valid for the selected suite unit 
Skipping package github.com/dgraph-io/dgraph/systest/bgindex as its not valid for the selected suite unit 
Skipping package github.com/dgraph-io/dgraph/systest/bulk_live/bulk as its not valid for the selected suite unit 
Skipping package github.com/dgraph-io/dgraph/systest/bulk_live/live as its not valid for the selected suite unit 
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/alpha isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/alpha/mutations_mode isCommon:false
Found valid task: github.com/dgraph-io/dgraph/ee/acl isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/admin_auth/poorman_auth isCommon:false
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/admin_auth/poorman_auth_with_acl isCommon:false
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/auth isCommon:false
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/auth/debug_off isCommon:false
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/auth_closed_by_default isCommon:false
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/common isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/custom_logic isCommon:false
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/custom_logic/cmd isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/directives isCommon:false
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/multi_tenancy isCommon:false
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/normal isCommon:false
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/schema isCommon:false
Found valid task: github.com/dgraph-io/dgraph/graphql/e2e/subscription isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest isCommon:true
Found valid task: github.com/dgraph-io/dgraph/systest/21million/common isCommon:true
Found valid task: github.com/dgraph-io/dgraph/systest/audit isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/backup/common isCommon:true
Found valid task: github.com/dgraph-io/dgraph/systest/backup/encryption isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/backup/filesystem isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/backup/minio isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/backup/minio-large isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/backup/multi-tenancy isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/bulk_live/common isCommon:true
Found valid task: github.com/dgraph-io/dgraph/systest/cdc isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/export isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/group-delete isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/license isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/loader isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/multi-tenancy isCommon:true
Found valid task: github.com/dgraph-io/dgraph/systest/online-restore isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/plugin isCommon:false
Found valid task: github.com/dgraph-io/dgraph/systest/posting-list-benchmark isCommon:true
Found valid task: github.com/dgraph-io/dgraph/worker isCommon:false
Found valid task: github.com/dgraph-io/dgraph/ee/audit isCommon:true
Found valid task: github.com/dgraph-io/dgraph/ee/backup isCommon:true
Found valid task: github.com/dgraph-io/dgraph/ee/enc isCommon:true
Found valid task: github.com/dgraph-io/dgraph/filestore isCommon:true
Found valid task: github.com/dgraph-io/dgraph/gql isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/admin isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/api isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/authorization isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/bench isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/dgraph isCommon:true
Found valid task: github.com/dgraph-io/dgraph/compose isCommon:true
Found valid task: github.com/dgraph-io/dgraph/conn isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/embargo isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/integration/acctupsert isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/integration/bank isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/integration/bigdata isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/integration/mutates isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/integration/swap isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/integration/testtxn isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/jepsen isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/jepsen/browser isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/neo4j-converter isCommon:true
Found valid task: github.com/dgraph-io/dgraph/contrib/teamcity isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/resolve isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/schema isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/subscription isCommon:true
Found valid task: github.com/dgraph-io/dgraph/graphql/test isCommon:true
Found valid task: github.com/dgraph-io/dgraph/lex isCommon:true
Found valid task: github.com/dgraph-io/dgraph/posting isCommon:true
Found valid task: github.com/dgraph-io/dgraph/protos isCommon:true
Found valid task: github.com/dgraph-io/dgraph/protos/pb isCommon:true
Found valid task: github.com/dgraph-io/dgraph/query isCommon:true
Found valid task: github.com/dgraph-io/dgraph/raftwal isCommon:true
Found valid task: github.com/dgraph-io/dgraph/schema isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph isCommon:false
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd isCommon:true
Found valid task: github.com/dgraph-io/dgraph/algo isCommon:true
Found valid task: github.com/dgraph-io/dgraph/chunker isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/bulk isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/cert isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/conv isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/debug isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/debuginfo isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/decrypt isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/increment isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/live isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/live/load-json isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/live/load-uids isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/migrate isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/version isCommon:true
Found valid task: github.com/dgraph-io/dgraph/dgraph/cmd/zero isCommon:true
Found valid task: github.com/dgraph-io/dgraph/edgraph isCommon:true
Found valid task: github.com/dgraph-io/dgraph/ee isCommon:true
Found valid task: github.com/dgraph-io/dgraph/t isCommon:true
Found valid task: github.com/dgraph-io/dgraph/task isCommon:true
Found valid task: github.com/dgraph-io/dgraph/telemetry isCommon:true
Found valid task: github.com/dgraph-io/dgraph/testutil isCommon:true
Found valid task: github.com/dgraph-io/dgraph/testutil/custom_plugins/anagram isCommon:true
Found valid task: github.com/dgraph-io/dgraph/testutil/custom_plugins/cidr isCommon:true
Found valid task: github.com/dgraph-io/dgraph/testutil/custom_plugins/factor isCommon:true
Found valid task: github.com/dgraph-io/dgraph/testutil/custom_plugins/rune isCommon:true
Found valid task: github.com/dgraph-io/dgraph/tlstest/acl isCommon:false
Found valid task: github.com/dgraph-io/dgraph/tlstest/certrequest isCommon:false
Found valid task: github.com/dgraph-io/dgraph/tlstest/certrequireandverify isCommon:false
Found valid task: github.com/dgraph-io/dgraph/tlstest/certverifyifgiven isCommon:false
Found valid task: github.com/dgraph-io/dgraph/tlstest/mtls_internal/ha_6_node isCommon:false
Found valid task: github.com/dgraph-io/dgraph/tlstest/mtls_internal/multi_group isCommon:false
Found valid task: github.com/dgraph-io/dgraph/tlstest/mtls_internal/single_node isCommon:false
Found valid task: github.com/dgraph-io/dgraph/tlstest/zero_https/all_routes_tls isCommon:false
Found valid task: github.com/dgraph-io/dgraph/tlstest/zero_https/no_tls isCommon:false
Found valid task: github.com/dgraph-io/dgraph/tok isCommon:true
Found valid task: github.com/dgraph-io/dgraph/types isCommon:true
Found valid task: github.com/dgraph-io/dgraph/types/facets isCommon:true
Found valid task: github.com/dgraph-io/dgraph/upgrade isCommon:true
Found valid task: github.com/dgraph-io/dgraph/codec isCommon:true
Found valid task: github.com/dgraph-io/dgraph/x isCommon:true
Found valid task: github.com/dgraph-io/dgraph/xidmap isCommon:true
Running tests for 114 packages.
Sent 1/114 packages for processing.
Sent 2/114 packages for processing.
Sent 3/114 packages for processing.
Bringing up cluster for package: github.com/dgraph-io/dgraph/dgraph/cmd/alpha/mutations_mode
Bringing up cluster test-992-3 for package: ../dgraph/docker-compose.yml ...
Bringing up cluster test-992-4 for package: ../dgraph/cmd/alpha/mutations_mode/docker-compose.yml ...
Bringing up cluster test-992-1 for package: ../dgraph/docker-compose.yml ...
While running command: "docker-compose -f ../dgraph/cmd/alpha/mutations_mode/docker-compose.yml -p test-992-4 up --force-recreate --build --remove-orphans --detach" Error: exit status 1
While running command: "docker-compose -f ../dgraph/docker-compose.yml -p test-992-3 up --force-recreate --build --remove-orphans --detach" Error: exit status 1
Running: /usr/bin/go test -failfast -v -timeout 30m github.com/dgraph-io/dgraph/ee/acl with test-992-3
While running command: "docker-compose -f ../dgraph/docker-compose.yml -p test-992-1 up --force-recreate --build --remove-orphans --detach" Error: exit status 1
Running: /usr/bin/go test -failfast -v -timeout 30m github.com/dgraph-io/dgraph/dgraph/cmd/alpha with test-992-1
CLUSTER DOWN: test-992-2
Docker logs for 920caef178652bd66a2bf1fb5d7eb1ccbe264fa6e621d52fb60938da2a1adeea is  with error <nil> Docker logs for 20b6a230eb352e3f96da8ff3deb80a65775c6b958d220859955efdee18a8e2bc is  with error <nil> Docker logs for e4db5f70782008a84da392ff6f52cf6078cffb26c2dcb2657342086e96b9c839 is  with error <nil> Docker logs for ba22ea8565f359669552332bcb8b2dcd35744ef44b9002abdbd3ce78c15fb4c2 is  with error <nil> CLUSTER DOWN: test-992-1
CLUSTER DOWN: test-992-3
TIMELINE starting at 10:57:02 PM
[    3s][0] COMPILE took: 3s
[   15s]   [1] DONE took: 10s
[   17s]      [2] DONE took: 13s
[   17s]         [3] DONE took: 13s
Got error: exit status 1.
Tests FAILED.

  ```
</details>


<details>
  <summary>Output of `./t -s -p "query"`</summary>

```log
testutil: "" localhost:9080 localhost:5080
Proc ID is 674
Installing dgraph ...
make[1]: Entering directory '/home/samyak/PESU/Sem6/HP/Project/dgraph/dgraph'
Commit SHA256: a44abf8587f4c26714952c45a05aedff1373ab13
Old SHA256: 561df6fcce9beff32df455feed232cf4b9d1a726fed7a4ca93abf44613ab7260
New SHA256: 561df6fcce9beff32df455feed232cf4b9d1a726fed7a4ca93abf44613ab7260
make[1]: Leaving directory '/home/samyak/PESU/Sem6/HP/Project/dgraph/dgraph'
compiling plugin: src="./custom_plugins/anagram/main.go" so="./custom_plugins/0.so"
compiling plugin: src="./custom_plugins/cidr/main.go" so="./custom_plugins/1.so"
compiling plugin: src="./custom_plugins/factor/main.go" so="./custom_plugins/2.so"
compiling plugin: src="./custom_plugins/rune/main.go" so="./custom_plugins/3.so"
plugin build completed. Files are: /home/samyak/PESU/Sem6/HP/Project/dgraph/t/custom_plugins/0.so,/home/samyak/PESU/Sem6/HP/Project/dgraph/t/custom_plugins/1.so,/home/samyak/PESU/Sem6/HP/Project/dgraph/t/custom_plugins/2.so,/home/samyak/PESU/Sem6/HP/Project/dgraph/t/custom_plugins/3.so
Found valid task: github.com/dgraph-io/dgraph/query isCommon:true
Running tests for 1 packages.
Sent 1/1 packages for processing.
Bringing up cluster test-674-1 for package: ../dgraph/docker-compose.yml ...
While running command: "docker-compose -f ../dgraph/docker-compose.yml -p test-674-1 up --force-recreate --build --remove-orphans --detach" Error: exit status 1
Running: /usr/bin/go test -failfast -v -timeout 30m github.com/dgraph-io/dgraph/query with test-674-1
Docker logs for 40d1a1d2875351cd217fabb25f8ecdd97217ac88d469bb2f12a702e2a50b138d is  with error <nil> Docker logs for 4c6d3d33ac7be1e98813fe6decf32e5c6f3ebc0b45e385dc8cf02acd6bf8d09b is  with error <nil> CLUSTER DOWN: test-674-1
TIMELINE starting at 10:58:11 PM
[    2s][0] COMPILE took: 2s
[   10s]   [1] DONE took: 7s
Got error: While running command: [go test -failfast -v -timeout 30m github.com/dgraph-io/dgraph/query] Error: context canceled.
Tests FAILED.

```
</details>

I tried rebasing on master, the error persists.